+++ machop/chop/actions/search/strategies/__init__.py
@@ -1,7 +1,6 @@
-# from .rl import StrategyRL
+from .rl import SearchStrategyRL
 from .optuna import SearchStrategyOptuna
 from .base import SearchStrategyBase
-from chop.actions.search.strategies.TXLRLSearchStrategy import SearchStrategyRL
 
 
 SEARCH_STRATEGY_MAP = {
