+++ machop/chop/actions/search/strategies/rl/__init__.py
@@ -1,1 +1,1 @@
-from .core_algorithm import StrategyRL
+from .core_algorithm import SearchStrategyRL
