+++ txl_workplace/rl.toml
@@ -4,13 +4,14 @@ dataset = "cifar10"
 task = "cls"
 
 max_epochs = 300
-batch_size = 1024
+batch_size = 512
 learning_rate = 1e-2
 accelerator = "cpu"
 project = "rl_search"
 seed = 42
 log_every_n_steps = 5
-load_name = "/mnt/c/workplace/projects/ADLS/mase/txl_workplace/test-accu-0.9332.ckpt"
+#load_name = "/mnt/c/workplace/projects/ADLS/mase/txl_workplace/test-accu-0.9332.ckpt"
+load_name = "/mnt/d/project_files/Python/mase/txl_workplace/test-accu-0.9332.ckpt"
 load_type = "pl"
 
 [passes.quantize]
@@ -39,20 +40,20 @@ name = ["NA"]
 [search.search_space.seed.linear.config]
 # if search.search_space.setup.by = "type", this seed will be used to quantize all torch.nn.Linear/ F.linear
 name = ["integer"]
-data_in_width = [6, 7, 8, 9]
+data_in_width = [6, 8, 10, 12, 14, 16]
 data_in_frac_width = ["NA"] # "NA" means data_in_frac_width = data_in_width // 2
-weight_width = [5, 6, 7]
+weight_width = [6, 8, 10, 12, 14, 16]
 weight_frac_width = ["NA"]
-bias_width = [5, 6, 7, 8]
+bias_width = [6, 8, 10, 12, 14, 16]
 bias_frac_width = ["NA"]
 
 [search.search_space.seed.conv2d.config]
 name = ["integer"]
-data_in_width = [8, 9, 10, 11, 12, 14]
+data_in_width = [6, 8, 10, 12, 14, 16]
 data_in_frac_width = ["NA"] # "NA" means data_in_frac_width = data_in_width // 2
-weight_width = [4, 6, 7, 8]
+weight_width = [6, 8, 10, 12, 14, 16]
 weight_frac_width = ["NA"]
-bias_width = [4, 5, 6, 7, 8, 10]
+bias_width = [6, 8, 10, 12, 14, 16]
 bias_frac_width = ["NA"]
 
 #[search.search_space.seed.relu.config]
@@ -73,10 +74,10 @@ compare_to = 32 # compare to FP32
 
 [search.strategy.setup]
 device = "cpu"
-total_trials = 100
+total_trials = 1
 algorithm = "a2c"
 env = "MixedPrecisionEnv"
 
 [search.strategy.metrics]
-accuracy.scale = 10.0
-average_bitwidth.scale = -0.1
+accuracy.scale = 1.0
+average_bitwidth.scale = -0.03125
